requirements.txt
streamlit
pillow  # (If you're using Version 3 with the image)